# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Moj Laravel API",
        "description": "Ovo je dokumentacija za moj Laravel API projekat",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni razvojni server"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Registracija korisnika",
                "description": "Kreira novi korisnički nalog i vraća token za autentifikaciju.",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Marko Markovic"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marko@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "lozinka123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspesna registracija",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdef123456..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Greska validacije"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Prijava korisnika",
                "description": "Proverava kredencijale i generise token za autentifikaciju.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marko@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "lozinka123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspesna prijava",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdef123456..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Pogresni kredencijali"
                    },
                    "422": {
                        "description": "Greska validacije"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Odjava korisnika",
                "description": "Brise trenutni token i odjavljuje korisnika.",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Uspesna odjava"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/change-password": {
            "put": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Promena lozinke",
                "description": "Menja lozinku ulogovanog korisnika.",
                "operationId": "c9cf137ca229152d78da153d798cb251",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "stara123"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "nova123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lozinka uspesno promenjena"
                    },
                    "401": {
                        "description": "Pogresna trenutna lozinka"
                    },
                    "422": {
                        "description": "Greska validacije"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/forgot-password": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Zahtev za reset lozinke",
                "description": "Generise token za resetovanje lozinke.",
                "operationId": "7af8817e88ba5874543ff38f188248f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marko@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token uspesno generisan"
                    },
                    "422": {
                        "description": "Greska validacije"
                    }
                }
            }
        },
        "/api/reset-password": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Resetovanje lozinke",
                "description": "Resetuje lozinku korisnika pomocu tokena.",
                "operationId": "00c0aad74830221c6da409a1355ec736",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marko@gmail.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "abc123..."
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "nova123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lozinka uspesno resetovana"
                    },
                    "404": {
                        "description": "Korisnik nije pronadjen"
                    },
                    "422": {
                        "description": "Greska validacije"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Autentifikacija",
            "description": "Autentifikacija"
        }
    ]
}